+++ packages/ozone/src/setting/validators.ts
@@ -1,6 +1,6 @@
 import { Selectable } from 'kysely'
 import { Setting } from '../db/schema/setting'
-import { ProtectedTagSettingKey } from './constants'
+import { PolicyListSettingKey, ProtectedTagSettingKey } from './constants'
 import { InvalidRequestError } from '@atproto/xrpc-server'
 
 export const settingValidators = new Map<
@@ -54,6 +54,33 @@ export const settingValidators = new Map<
               `Must define at least one moderator DID for tag ${key}`,
             )
           }
+        }
+      }
+    },
+  ],
+  [
+    PolicyListSettingKey,
+    async (setting: Partial<Selectable<Setting>>) => {
+      if (setting.managerRole !== 'tools.ozone.team.defs#roleAdmin') {
+        throw new InvalidRequestError(
+          'Only admins should be able to manage policy list',
+        )
+      }
+
+      if (typeof setting.value !== 'object') {
+        throw new InvalidRequestError('Invalid value')
+      }
+      for (const [key, val] of Object.entries(setting.value)) {
+        if (!val || typeof val !== 'object') {
+          throw new InvalidRequestError(
+            `Invalid configuration for policy ${key}`,
+          )
+        }
+
+        if (!val['name'] || !val['description']) {
+          throw new InvalidRequestError(
+            `Must define a name and description for policy ${key}`,
+          )
         }
       }
     },