+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atproto/api
 
+## 0.13.25
+
+### Patch Changes
+
+- [#3271](https://github.com/bluesky-social/atproto/pull/3271) [`53621f8e1`](https://github.com/bluesky-social/atproto/commit/53621f8e100a3aa3c1caff10a08d3f4ea919875a) Thanks [@foysalit](https://github.com/foysalit)! - Allow setting policy names with takedown actions and when querying events
+
 ## 0.13.24
 
 ### Patch Changes